js: Web Crypto support for AES-CMAC
This commit adds a WebCryptoAesCmac implementation as an alternative to
PolyfillAesCmac, which is built using an AES-ECB-like wrapper on top of CBC mode
(i.e. use an IV of all zeroes and take the first block of CBC mode output)

This could be further refined to use CBC mode to process multiple blocks instead
of just one, which should improve performance. This commit is just a first pass.