js: Refactor in preparation for AEAD API
The previous package structure made it somewhat difficult to expose
additional classes. Really `AEAD` and `SIV` should sit side-by-side.

This removes the `Miscreant` class and makes `SIV` a first-class API in
anticipation of adding `AEAD` (and also `STREAM`).

`CMAC` and `PMAC` are now capitalized and located in the `mac`
directory.

Providers have been moved to a `providers` directory.

The `internal` directory was renamed to `internals`.