Remove WebCrypto tests; "polyfill" -> "soft"; fix CI
The `node-webcrypto-ossl` that was previously used for testing the
WebCrypto backend seems unmaintained and is uninstallable now.

This removes all tests for it: we should migrate from WebCrypto to a
WASM build of the Rust implementation or Node binding to the native
binary instead.

It also renames "Polyfill" to "Soft" and updates various copyright
dates.